+++ DailyReport/js/app.js
@@ -74,7 +74,7 @@ function startPage() {
   sequenceValidation();
   initCalendar();
   getPlans();
-  // hideOvertimeButton();
+  hideOvertimeButton();
 
   $(".cs-loader").fadeOut(1000);
   const btn = document.getElementById("btnRequestAccess");