+++ Override/php/delete_entry.php
@@ -1,9 +1,14 @@
 <?php
 #region DB Connect
 require_once "../../dbconn/dbconnectwebjmr.php";
+require_once "./dr_history.php";
 #endregion
 
 #region Initialize Variable
+$result = [
+  "isSuccess" => FALSE,
+  "message" => ''
+];
 if(!empty($_POST['drID'])){
     $drID = $_POST['drID'];
 } else {
@@ -15,10 +20,16 @@ if(!empty($_POST['drID'])){
 
 #region Main Query
 try {
+  $entryId = (int) $drID;
+  overrideLoadDailyReportHistory();
+  $oldRow = fetchDailyReportRowById($entryId) ?: [];
+  $actorNum = overrideHistoryActorNum($_POST);
+
   $deleteQ = "DELETE FROM dailyreport WHERE fldID = :drID";
   $deleteStmt = $connwebjmr->prepare($deleteQ);
   $deleteStmt->execute([":drID"=>$drID]);
   if($deleteStmt->rowCount() > 0) {
+    overrideRecordDeletedHistory($oldRow, $entryId, $actorNum);
     $result['isSuccess'] = TRUE;
     $result['message'] = "Successfully Deleted!";
   }